Received Kavenegar

+++ SmsHub.Persistance/Migrations/1403_06_15_DbInitialDesign.cs
@@ -391,7 +391,7 @@ namespace SmsHub.Persistence.Migrations
         private void CreateReceive()
         {
             Create.Table(nameof(TableName.Received))
-                .WithColumn(Id).AsInt32().PrimaryKey().Identity()
+                .WithColumn(Id).AsInt64().PrimaryKey().Identity()
                 .WithColumn("MessageId").AsInt64().Nullable()
                 .WithColumn("Status").AsInt32().Nullable()
                 .WithColumn("MessageText").AsString(int.MaxValue)
